refactor : status enum을 card 안으로 이동
- enum을 이동 시켰다.
- 일급 컬렉션인 logs를 사용한다.
