--- backend/src/main/java/com/codeSquad/cocokyu/domain/card/Status.java
@@ -1,8 +1,0 @@
-package com.codeSquad.cocokyu.domain.card;
-
-public enum Status {
-    TODO,
-    DOING,
-    DONE,
-    DELETED
-}
