+++ backend/src/main/java/com/codeSquad/cocokyu/domain/model/Card.java
@@ -1,4 +1,4 @@
-package com.codeSquad.cocokyu.domain.card;
+package com.codeSquad.cocokyu.domain.model;
 
 import org.springframework.data.annotation.Id;
 
@@ -13,11 +13,17 @@ public class Card {
     private LocalDateTime createDateTime;
     private Status status;
 
+    private Logs logs = new Logs();
+
+    protected Card() {
+    }
+
     public Card(String title, String contents, Status status) {
         this.title = title;
         this.contents = contents;
         this.status = status;
         this.createDateTime = LocalDateTime.now();
+        logs.createLog(this);
     }
 
     public Long getId() {
@@ -41,13 +47,21 @@ public class Card {
     }
 
     public void modify(Card updateCard) {
+        logs.updateLog(this, updateCard);
         this.title = updateCard.title;
         this.contents = updateCard.contents;
         this.status = updateCard.status;
     }
 
-    public void changeStatus(Status status){
-        this.status = status;
+    public void delete() {
+        logs.deleteLog(this);
+        this.status = Status.DELETED;
     }
 
+    public enum Status {
+        TODO,
+        DOING,
+        DONE,
+        DELETED
+    }
 }
